Copy busybox rather than symlinking it
Install busybox by copying the executable and not symlinking it. The
search for duplicates will automatically hardlink together.
Hardlinks makes more sense in an environment like this.

Change the test to avoid creating directories in DESTDIR. Copying
busybox to DESTDIR/linuxrc fail because of the test since the directory
doesn't exists. copyfile() don't check when the file is directly in
DESTDIR.

Signed-off-by: Louis Leseur <louis.leseur@gmail.com>