Add support for custom binary root directory
Use --binroot /path/to/root to use binaries from a system installed in a
given root directory.

Fix: findexec() check file is executable.
Fix: findlib, findexec and find_elf_deps does not check for binary
compatibility when the given path is absolute.
Change: find*() only check in the root directory when the given path is
absolute.